+++ R/set-spawning-habitat.R
@@ -131,7 +131,7 @@ SR_spawn_approx <- function(relationship_df, modeling_lookup){
     # case when modeling exists
     SR_approx <- approxfun(relationship_df$flow_cfs, relationship_df$SR_spawn_wua, rule = 2)
   } else {
-    # case when no spring run modeling but fall run modeling is used
+    # use fall run if availble, defaults to fall if not
     SR_approx <- FR_spawn_approx(relationship_df, modeling_lookup)
   }