+++ NEWS.md
@@ -5,5 +5,7 @@ Adjust Mokelumne Spawning Data 10/30/2018
 # cvpiaHabitat 1.0.2 (in progress)
 Adjust Stanislaus rearing habitat for steelhead
 Add vignette describing floodplain habitat scaling
-Mokelumne
-American spawning WUA updated
+Mokelumne update instream and floodplain rearing habitat
+American spawning WUA update
+Clear creek spawning update
+